Refactor HSTU attention under shared namespace
