+++ benchmark/hstu/hstu_attention/.gitignore
@@ -1,0 +1,2 @@
+# Local benchmark analysis scripts and generated results.
+/.artifacts/
