+++ python/cudnn/__init__.py
@@ -361,10 +361,10 @@ _LAZY_OPTIONAL_IMPORTS = {
     "grouped_gemm_srelu_wrapper_sm100": (".gemm.cutedsl.grouped", "grouped_gemm_srelu_wrapper_sm100"),
     "GroupedGemmDsreluSm100": (".gemm.cutedsl.grouped", "GroupedGemmDsreluSm100"),
     "grouped_gemm_dsrelu_wrapper_sm100": (".gemm.cutedsl.grouped", "grouped_gemm_dsrelu_wrapper_sm100"),
-    "HSTUFwdSm100": (".hstu_attention", "HSTUFwdSm100"),
-    "HSTUBwdSm100": (".hstu_attention", "HSTUBwdSm100"),
-    "hstu_attention_forward": (".hstu_attention", "hstu_attention_forward"),
-    "hstu_attention_backward": (".hstu_attention", "hstu_attention_backward"),
+    "HSTUFwdSm100": (".hstu.hstu_attention", "HSTUFwdSm100"),
+    "HSTUBwdSm100": (".hstu.hstu_attention", "HSTUBwdSm100"),
+    "hstu_attention_forward": (".hstu.hstu_attention", "hstu_attention_forward"),
+    "hstu_attention_backward": (".hstu.hstu_attention", "hstu_attention_backward"),
     "GroupedGemmQuantSm100": (".gemm.cutedsl.grouped", "GroupedGemmQuantSm100"),
     "grouped_gemm_quant_wrapper_sm100": (".gemm.cutedsl.grouped", "grouped_gemm_quant_wrapper_sm100"),
     "GroupedGemmGluSm100": (".gemm.cutedsl.grouped", "GroupedGemmGluSm100"),
