+++ test/python/fe_api/hstu/hstu_attention/test_hstu_attention.py
@@ -19,13 +19,13 @@ try:
 except (ImportError, OSError) as exc:
     pytest.skip(f"CuTe DSL is unavailable: {exc}", allow_module_level=True)
 
-from cudnn.hstu_attention import (
+from cudnn.hstu.hstu_attention import (
     HSTUBwdSm100,
     HSTUFwdSm100,
     hstu_attention_backward,
     hstu_attention_forward,
 )
-from cudnn.hstu_attention import _interface, api as _api
+from cudnn.hstu.hstu_attention import _interface, api as _api
 
 pytestmark = [
     pytest.mark.gpu_exclusive,
@@ -1288,7 +1288,7 @@ def test_explicit_api_reuses_direct_grad_kernel_for_aligned_strides(monkeypatch)
 @pytest.mark.L0
 @pytest.mark.skipif(not _IS_SM10X, reason="requires an SM10x Blackwell GPU")
 def test_d256_explicit_compile_and_packed_gradient_outputs():
-    from cudnn.hstu_attention._kernels.hstu_bwd_256_cute import (
+    from cudnn.hstu.hstu_attention._kernels.hstu_bwd_256_cute import (
         hstu_varlen_bwd_256_cute,
     )
 
@@ -2301,7 +2301,7 @@ def test_runtime_alpha_and_scaling_are_not_compile_time_constants(head_dim):
         assert len(_interface.hstu_varlen_fwd_100.compile_cache) == 1
 
     if head_dim == 256:
-        from cudnn.hstu_attention._kernels.hstu_bwd_256_cute import (
+        from cudnn.hstu.hstu_attention._kernels.hstu_bwd_256_cute import (
             hstu_varlen_bwd_256_cute,
         )
